Convert the TLS version enum to an enum class (GH #3146)

+++ src/lib/tls/tls_version.cpp
@@ -84,11 +84,11 @@ bool Protocol_Version::valid() const
 
 bool Protocol_Version::known_version() const
    {
-   return (m_version == Protocol_Version::TLS_V12 ||
+   return (m_version == static_cast<uint16_t>(Protocol_Version::TLS_V12) ||
 #if defined(BOTAN_HAS_TLS_13)
-           m_version == Protocol_Version::TLS_V13 ||
+           m_version == static_cast<uint16_t>(Protocol_Version::TLS_V13) ||
 #endif
-           m_version == Protocol_Version::DTLS_V12);
+           m_version == static_cast<uint16_t>(Protocol_Version::DTLS_V12));
    }
 
 }
